updated services so there aren't any responsentity interactions/classes. Controller should only care about Http stuffs.

+++ beer-catalog-service/src/main/java/com/evolutionaryeyes/beerproductservice/dto/BeerItemDTO.java
@@ -15,8 +15,6 @@ public class BeerItemDTO {
     private String description;
     private String brewery;
     private Double abv;
-    private Double ibu;
+    private int ibu;
     private String type;
-
-
 }
